requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
colorama==0.4.5
idna==3.3
interchange==2021.0.4
monotonic==1.6
numpy==1.21.6
packaging==21.3
pandas==1.3.5
pansi==2020.7.3
protobuf==3.20.0
py2neo==2021.2.3
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
stanford-openie==1.3.0
stanfordnlp==0.2.0
torch==1.12.0
tqdm==4.64.0
typing-extensions==4.3.0
urllib3==1.26.9
wget==3.2
